backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
torch>=2.6.0
torchvision>=0.17.0
numpy>=1.24.3
Pillow>=10.0.1
opencv-python>=4.8.1.78
trimesh>=4.0.5
pyglet>=2.0.9
psutil>=5.9.5
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Database - SQLite (simpler than PostgreSQL)
sqlalchemy>=2.0.23
alembic>=1.13.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0